# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Akmatori API
  description: |
    AI-powered AIOps platform API. Manages skills, tools, incidents, alert sources,
    and system settings for automated incident investigation and remediation.
  version: 1.0.0
  contact:
    name: Akmatori
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors
...
    Integration:
      type: object
      description: |
        A configured connection to a messaging provider (Slack, Telegram, ...).
        Credentials are stored as a JSON blob whose shape is provider-specific.
      properties:
        id: {type: integer}
        uuid: {type: string, format: uuid}
        provider:
          type: string
          enum: [slack, telegram]
        name: {type: string}
        credentials:
          type: object
          additionalProperties: true
        enabled: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    IntegrationCreateRequest:
      type: object
      required: [provider, name]
      properties:
        provider:
          type: string
          enum: [slack, telegram]
        name: {type: string}
        credentials:
          type: object
          additionalProperties: true
        enabled: {type: boolean}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /integrations:
    get:
      summary: List messaging provider integrations
      operationId: listIntegrations
      tags: [Integrations]
      responses:
        '200':
          description: Configured integrations
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Integration'}
        '503':
          description: Channel service is not configured
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    post:
      summary: Create a new integration
      operationId: createIntegration
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IntegrationCreateRequest'}
      responses:
        '201':
          description: Integration created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Integration'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflict (duplicate name or invariant violation)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...